# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: |
    This is the OpenRQM API specification.
    [OpenRQM Docs - Github](https://github.com/openrqm/openrqm-docs)
  license:
    name: LGPL-2.0-only
    url: https://spdx.org/licenses/LGPL-2.0-only.html#licenseText
  title: OpenRQM
  version: 1.0.1
...
servers:
- url: https://virtserver.swaggerhub.com/OpenRQM/OpenRQM/1.0.0
security:
- bearerAuth: []
...
paths:
...
  /workspace:
    delete:
      operationId: deleteWorkspace
      parameters:
      - description: The workspace to delete
        explode: true
        in: query
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          description: The workspace was deleted successfully
      summary: Delete a workspace
      tags:
      - Workspaces
      x-accepts: application/json
      x-tags:
      - tag: Workspaces
    get:
      operationId: getWorkspace
      parameters:
      - description: The workspace identifier
        explode: true
        in: query
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RQMWorkspace'
          description: The workspace was fetched successfully
      summary: Get a workspace
      tags:
      - Workspaces
      x-accepts: '*/*'
      x-tags:
      - tag: Workspaces
    patch:
      operationId: patchWorkspace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RQMWorkspace'
        description: The workspace to update
        required: true
      responses:
        "200":
          description: Updating workspace successful
      summary: Update a workspace
      tags:
      - Workspaces
      x-contentType: application/json
      x-accepts: application/json
      x-tags:
      - tag: Workspaces
    post:
      operationId: postWorkspace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RQMWorkspace'
        description: The workspace to create
        required: true
      responses:
        "200":
          description: The workspace was created by the backend
      summary: Add a workspace
      tags:
      - Workspaces
      x-contentType: application/json
      x-accepts: application/json
      x-tags:
      - tag: Workspaces
...
  /export/openrqm:
    get:
      operationId: exportRaw
      parameters:
      - description: The document to export
        explode: true
        in: query
        name: documentId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            application/zip:
              schema:
                format: binary
                type: string
          description: Document successfully exported
      summary: Export a document as OpenRQM Data for import
      tags:
      - Export
      x-accepts: application/zip
      x-tags:
      - tag: Export
  /export/pdf:
    get:
      operationId: exportPdf
      parameters:
      - description: The document to export
        explode: true
        in: query
        name: documentId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      - description: The template to use for the export
        explode: true
        in: query
        name: templateId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            application/pdf:
              schema:
                format: binary
                type: string
          description: Document successfully exported
      summary: Export a document as PDF
      tags:
      - Export
      x-accepts: application/pdf
      x-tags:
      - tag: Export
...
  /export/markdown:
    get:
      operationId: exportMarkdown
      parameters:
      - description: The document to export
        explode: true
        in: query
        name: documentId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      - description: The template to use for the export
        explode: true
        in: query
        name: templateId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            application/zip:
              schema:
                format: binary
                type: string
          description: Document successfully exported
      summary: Export a document as Markdown
      tags:
      - Export
      x-accepts: application/zip
      x-tags:
      - tag: Export
...
  /theme:
    get:
      operationId: getTheme
      parameters:
      - description: The theme id
        explode: true
        in: query
        name: themeId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RQMTheme'
          description: Fetching theme successful
      summary: Get a theme
      tags:
      - Themes
      x-accepts: '*/*'
      x-tags:
      - tag: Themes
    patch:
      operationId: updateTheme
      requestBody:
        $ref: '#/components/requestBodies/RQMTheme'
      responses:
        "200":
          description: Updating theme successful
      summary: Update a theme
      tags:
      - Themes
      x-contentType: application/json
      x-accepts: application/json
      x-tags:
      - tag: Themes
    post:
      operationId: createTheme
      requestBody:
        $ref: '#/components/requestBodies/RQMTheme'
      responses:
        "200":
          description: Creating theme successful
      summary: Create a new theme
      tags:
      - Themes
      x-contentType: application/json
      x-accepts: application/json
      x-tags:
      - tag: Themes
  /theme/elementTypeColors:
    get:
      operationId: getElementTypeColorsOfTheme
      parameters:
      - explode: true
        in: query
        name: themeId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/RQMElementTypeColor'
                type: array
          description: Get ElementTypeColors successful
      summary: Get all element type colors of a theme
      tags:
      - Themes
      x-accepts: '*/*'
      x-tags:
      - tag: Themes
...
  /link:
    delete:
      operationId: deleteLink
      parameters:
      - description: The id of the link
        explode: true
        in: query
        name: linkId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          description: Deleting link was successful
      summary: Delete a link
      tags:
      - Links
      x-accepts: application/json
      x-tags:
      - tag: Links
    post:
      operationId: linkElement
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RQMLink'
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RQMLink'
          description: Linking was successful
      summary: Link two elements
      tags:
      - Links
      x-contentType: application/json
      x-accepts: '*/*'
      x-tags:
      - tag: Links
...
components:
  requestBodies:
    RQMTheme:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RQMTheme'
      required: true
...
  schemas:
    RQMWorkspace:
      example:
        documents:
        - approverId: 9
          confidentiality: confidentiality
          languageId: 7
          description: description
          previousBaselineId: 1
          authorId: 2
          baselineMinor: 4
          baselineMajor: 2
          lastModifiedOn: 2000-01-23T04:56:07.000+00:00
          baselineReview: 7
          internalIdentifier: 5
          lastModifiedById: 3
          reviewerText: reviewerText
          name: name
          externalIdentifier: externalIdentifier
          id: 1
          shortName: shortName
          workspaceId: 5
        - approverId: 9
          confidentiality: confidentiality
          languageId: 7
          description: description
          previousBaselineId: 1
          authorId: 2
          baselineMinor: 4
          baselineMajor: 2
          lastModifiedOn: 2000-01-23T04:56:07.000+00:00
          baselineReview: 7
          internalIdentifier: 5
          lastModifiedById: 3
          reviewerText: reviewerText
          name: name
          externalIdentifier: externalIdentifier
          id: 1
          shortName: shortName
          workspaceId: 5
        name: name
        id: 0
        workspaces:
        - null
        - null
        workspaceId: 6
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
        workspaceId:
          format: int64
          type: integer
        workspaces:
          items:
            $ref: '#/components/schemas/RQMWorkspace'
          type: array
        documents:
          items:
            $ref: '#/components/schemas/RQMDocument'
          type: array
      type: object
    RQMDocument:
      example:
        approverId: 9
        confidentiality: confidentiality
        languageId: 7
        description: description
        previousBaselineId: 1
        authorId: 2
        baselineMinor: 4
        baselineMajor: 2
        lastModifiedOn: 2000-01-23T04:56:07.000+00:00
        baselineReview: 7
        internalIdentifier: 5
        lastModifiedById: 3
        reviewerText: reviewerText
        name: name
        externalIdentifier: externalIdentifier
        id: 1
        shortName: shortName
        workspaceId: 5
      properties:
        id:
          format: int64
          type: integer
        workspaceId:
          format: int64
          type: integer
        internalIdentifier:
          format: int64
          type: integer
        externalIdentifier:
          type: string
        name:
          type: string
        shortName:
          type: string
        description:
          type: string
        confidentiality:
          type: string
        authorId:
          format: int64
          type: integer
        languageId:
          format: int64
          type: integer
        approverId:
          format: int64
          type: integer
        reviewerText:
          type: string
        lastModifiedById:
          format: int64
          type: integer
        lastModifiedOn:
          format: date-time
          type: string
        baselineMajor:
          format: int64
          type: integer
        baselineMinor:
          format: int64
          type: integer
        baselineReview:
          format: int64
          type: integer
        previousBaselineId:
          format: int64
          type: integer
      type: object
...
    RQMLink:
      example:
        fromDocumentShortName: fromDocumentShortName
        linkTypeId: 2
        fromElementId: 6
        toDocumentShortName: toDocumentShortName
        toDocumentId: 5
        id: 0
        fromDocumentId: 1
        toElementId: 5
      properties:
        id:
          description: Id of the link
          format: int64
          type: integer
        fromElementId:
          format: int64
          type: integer
        fromDocumentId:
          format: int64
          type: integer
        fromDocumentShortName:
          type: string
        toElementId:
          format: int64
          type: integer
        toDocumentId:
          format: int64
          type: integer
        toDocumentShortName:
          type: string
        linkTypeId:
          format: int64
          type: integer
      type: object
...
    RQMTheme:
      example:
        linkFromHighlightColor: linkFromHighlightColor
        linkToHighlightColor: linkToHighlightColor
        id: 0
      properties:
        id:
          format: int64
          type: integer
        linkFromHighlightColor:
          type: string
        linkToHighlightColor:
          type: string
      type: object
    RQMElementTypeColor:
      example:
        themeId: 0
        elementTypeId: 6
        colorCodeHex: colorCodeHex
      properties:
        themeId:
          format: int64
          type: integer
        elementTypeId:
          format: int64
          type: integer
        colorCodeHex:
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http